[Feat]: 대댓글 조회 기능 구현

+++ src/main/java/com/tenten/linkhub/domain/space/service/CommentService.java
@@ -1,6 +1,5 @@
 package com.tenten.linkhub.domain.space.service;
 
-import com.tenten.linkhub.domain.auth.MemberDetails;
 import com.tenten.linkhub.domain.space.model.space.Comment;
 import com.tenten.linkhub.domain.space.model.space.Space;
 import com.tenten.linkhub.domain.space.repository.comment.CommentRepository;
@@ -8,14 +7,18 @@ import com.tenten.linkhub.domain.space.repository.comment.dto.CommentAndChildCom
 import com.tenten.linkhub.domain.space.repository.space.SpaceRepository;
 import com.tenten.linkhub.domain.space.service.dto.CommentUpdateRequest;
 import com.tenten.linkhub.domain.space.service.dto.comment.CommentAndChildCountResponses;
+import com.tenten.linkhub.domain.space.service.dto.comment.RepliesFindResponses;
 import com.tenten.linkhub.domain.space.service.dto.comment.ReplyCreateRequest;
 import com.tenten.linkhub.domain.space.service.dto.comment.RootCommentCreateRequest;
 import com.tenten.linkhub.domain.space.service.mapper.CommentMapper;
 import com.tenten.linkhub.global.exception.DataNotFoundException;
+import com.tenten.linkhub.global.exception.UnauthorizedAccessException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 public class CommentService {
@@ -80,12 +83,16 @@ public class CommentService {
         Comment comment = commentRepository.findById(request.commentId())
                 .orElseThrow(() -> new DataNotFoundException("수정할 댓글을 찾을 수 없습니다."));
 
+        if (!Objects.equals(comment.getMemberId(), request.memberId())) {
+            throw new UnauthorizedAccessException("댓글을 수정 할 권한이 없습니다.");
+        }
+
         Comment updatedComment = comment.updateComment(request.content());
 
         return updatedComment.getId();
     }
 
-    public Long deleteComment(Long spaceId, Long commentId, MemberDetails memberDetails) {
+    public Long deleteComment(Long spaceId, Long commentId, Long memberId) {
         Space space = spaceRepository.getById(spaceId);
 
         space.validateCommentAvailability();
@@ -93,8 +100,22 @@ public class CommentService {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new DataNotFoundException("수정할 댓글을 찾을 수 없습니다."));
 
+        if (!Objects.equals(comment.getMemberId(), memberId)) {
+            throw new UnauthorizedAccessException("댓글을 삭제 할 권한이 없습니다.");
+        }
+
         commentRepository.deleteById(commentId);
 
         return commentId;
+    }
+
+    public RepliesFindResponses findReplies(Long spaceId, Long commentId, Pageable pageable) {
+        Space space = spaceRepository.getById(spaceId);
+
+        space.validateCommentAvailability();
+
+        Slice<Comment> responses = commentRepository.findRepliesById(commentId, pageable);
+
+        return RepliesFindResponses.from(responses);
     }
 }
